+++ src/ai/backend/manager/models/gql_models/image.py
@@ -11,6 +11,7 @@ from typing import (
     List,
     Optional,
     Self,
+    cast,
     overload,
 )
 from uuid import UUID
@@ -29,11 +30,13 @@ from ai.backend.common.bgtask.bgtask import ProgressReporter
 from ai.backend.common.docker import ImageRef, KernelFeatures, LabelName
 from ai.backend.common.exception import UnknownImageReference
 from ai.backend.common.types import (
+    AgentId,
     DispatchResult,
     ImageAlias,
 )
 from ai.backend.logging import BraceStyleAdapter
 from ai.backend.manager.data.container_registry.types import ContainerRegistryData
+from ai.backend.manager.models.minilang import EnumFieldItem
 from ai.backend.manager.models.minilang.ordering import ColumnMapType, QueryOrderParser
 from ai.backend.manager.models.minilang.queryfilter import (
     FieldSpecType,
@@ -126,6 +129,9 @@ __all__ = (
 _queryfilter_fieldspec: FieldSpecType = {
     "id": ("id", None),
     "name": ("name", None),
+    "namespace": ("image", None),
+    "tag": ("tag", None),
+    "status": (EnumFieldItem("status", ImageStatus), None),
     "project": ("project", None),
     "image": ("image", None),
     "created_at": ("created_at", dtparse),
@@ -133,13 +139,16 @@ _queryfilter_fieldspec: FieldSpecType = {
     "registry_id": ("registry_id", None),
     "architecture": ("architecture", None),
     "is_local": ("is_local", None),
-    "type": ("session_type", ImageType),
+    "type": (EnumFieldItem("type", ImageType), None),
     "accelerators": ("accelerators", None),
 }
 
 _queryorder_colmap: ColumnMapType = {
     "id": ("id", None),
     "name": ("name", None),
+    "namespace": ("image", None),
+    "tag": ("tag", None),
+    "status": ("status", None),
     "project": ("project", None),
     "image": ("image", None),
     "created_at": ("created_at", None),
@@ -464,6 +473,44 @@ class ImageNode(graphene.ObjectType):
         ImagePermissionValueField,
         description=f"Added in 25.3.0. One of {[val.value for val in ImagePermission]}.",
     )
+    installed = graphene.Boolean(
+        description="Added in 25.11.0. Indicates if the image is installed on any Agent."
+    )
+
+    @property
+    def _canonical(self) -> str:
+        image_ref = ImageRef(
+            self.base_image_name,
+            self.project,
+            self.tag,
+            self.registry,
+            self.architecture,
+            self.is_local,
+        )
+        return image_ref.canonical
+
+    @classmethod
+    async def _batch_load_installed_agents(
+        cls, ctx: GraphQueryContext, full_names: Sequence[str]
+    ) -> list[set[AgentId]]:
+        async def pipe_builder(r: Redis) -> Pipeline:
+            pipe = r.pipeline()
+            for name in full_names:
+                pipe.smembers(name)
+            return pipe
+
+        result = await redis_helper.execute(ctx.redis_image, pipe_builder)
+        result = cast(list[set[bytes]], result)
+        return [{AgentId(agent_id.decode()) for agent_id in agents} for agents in result]
+
+    async def resolve_installed(self, info: graphene.ResolveInfo) -> bool:
+        graph_ctx: GraphQueryContext = info.context
+        loader = graph_ctx.dataloader_manager.get_loader_by_func(
+            graph_ctx, self._batch_load_installed_agents
+        )
+        agent_ids = await loader.load(self._canonical)
+        agent_ids = cast(Optional[set[AgentId]], agent_ids)
+        return agent_ids is not None and len(agent_ids) > 0
 
     @classmethod
     async def batch_load_by_name_and_arch(