+++ src/ai/backend/test/testcases/session/creation_failure_too_many_container.py
@@ -5,6 +5,7 @@ from ai.backend.test.contexts.auth import KeypairContext
 from ai.backend.test.contexts.client_session import ClientSessionContext
 from ai.backend.test.contexts.image import ImageContext
 from ai.backend.test.contexts.session import SessionContext
+from ai.backend.test.contexts.tester import TestSpecMetaContext
 from ai.backend.test.templates.template import TestCode
 from ai.backend.test.utils.exceptions import UnexpectedSuccess
 
@@ -15,7 +16,9 @@ class SessionCreationFailureTooManyContainer(TestCode):
         client_session = ClientSessionContext.current()
         image_dep = ImageContext.current()
         session_dep = SessionContext.current()
-        session_name = "test-session-creation-failure"
+        spec_meta = TestSpecMetaContext.current()
+        test_id = spec_meta.test_id
+        session_name = f"test_failure_{str(test_id)}"
 
         access_key = keypair.access_key
         result = await client_session.KeyPair(access_key).info([keypair_fields["resource_policy"]])